Move initApplication to shared space
